backend/requirement.txt
Django==3.2.15
djangorestframework==3.13.1
markdown==3.4.1
django-filter==22.1
djangorestframework-simplejwt==5.2.0